Update variable names across multiple files
Renamed several ambiguous or non-descriptive variables in various files, predominantly in test cases, for better clarity and understanding of the code. The core changes were to replace 'dt' with 'dateTime' and 'sv' with 'stringVal', thus giving precise context to their uses. No change in functionality is introduced by this commit.